feat: analyze the dataset features
add some analysis about the dataset such as identifying the anomalies
days that has a high volume of trips or low than usual.

add more time formats that helps in some analysis.

analyze the traffic based on the average speed in each day identify the
days with a traffic congestion.

remove the abnormal passenger count in the dataset such as trips has 9 or
8 passengers and consider them outliers.